Exploring ScalaCheck to do property-based testing on UUIDs
